+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Tu 26-07-2016
+Version 1.2
+- Added support for CWD and @chdir
+- Added support for custom separator (default: '/')
+- Added import & export features
+
 Su 24-07-2016
 Version 1.1
 - Items can't take an object-reserved name anymore